move resolveinjavahome to javautil
